+++ src/extension/background-script/CryptoServices/appendShareTarget.ts
@@ -1,7 +1,7 @@
 import * as Alpha39 from '../../../crypto/crypto-alpha-39'
 import * as Gun2 from '../../../network/gun/version.2'
 import { ProfileIdentifier, PostIVIdentifier, GroupIdentifier, constructPostRecipients } from '../../../database/type'
-import { prepareOthersKeyForEncryptionV39OrV38 } from '../prepareOthersKeyForEncryption'
+import { prepareOthersKeyForEncryptionV39OrV38 } from './prepareOthersKeyForEncryption'
 import { cryptoProviderTable } from './utils'
 import { updatePostDB, RecipientDetail } from '../../../database/post'
 import { getNetworkWorker } from '../../../social-network/worker'
